[DEVELOPMENT ENVIRONMENT] * NEW: The project tree popup menu now has the refresh action.
[GB.SIGNAL]
* NEW: New component to catch or ignore process signals.


git-svn-id: svn://localhost/gambas/trunk@3037 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/lib/db/main.c
@@ -589,7 +589,7 @@ char *DB_GetQuotedTable(DB_DRIVER *driver, DB_DATABASE *db, const char *table)
 	else
 	{
 		res = GB.TempString(NULL, len + 2);
-		sprintf(res, "%.*s.%s%s%s", point - table, table, quote, point + 1, quote);
+		sprintf(res, "%.*s.%s%s%s", (int)(point - table), table, quote, point + 1, quote);
 	}
 	
 	return res;
